requirements.txt
aenum==2.2.3
alabaster==0.7.12
asn1crypto==0.24.0
astroid==2.1.0
asttokens==1.1.13
attrs==19.3.0
automationhat==0.2.0
awesem==1.0.0
Babel==2.7.0
beautifulsoup4==4.7.1
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
Cap1xxx==0.1.3
certifi==2018.8.24
chardet==3.0.4
Click==7.0
colorama==0.3.7
colorzero==1.1
cookies==2.2.1
cryptography==2.6.1
cycler==0.10.0
docutils==0.14
drumhat==0.1.0
entrypoints==0.3
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==1.0.2
fourletterphat==0.1.0
future==0.18.2
gnuplotlib==0.31
gpiozero==1.5.1
html5lib==1.0.1
idna==2.6
imagesize==1.1.0
importlib-metadata==0.23
isort==4.3.4
itsdangerous==0.24
jedi==0.13.2
Jinja2==2.10
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
logilab-common==1.4.2
loguru==0.3.2
lxml==4.3.2
m2r==0.2.1
MarkupSafe==1.1.0
matplotlib==3.1.1
mccabe==0.6.1
microdotphat==0.2.1
mistune==0.8.4
more-itertools==7.2.0
mote==0.0.4
motephat==0.0.2
mypy==0.670
mypy-extensions==0.4.1
numpy==1.16.2
numpy-indexed==0.3.5
numpysane==0.18
oauthlib==2.1.0
olefile==0.46
packaging==19.2
pantilthat==0.0.7
parso==0.3.1
pgzero==1.2
phatbeat==0.1.1
pi-plates==5.22
pianohat==0.1.0
picamera==1.13
picraft==1.0
piglow==1.2.5
pigpio==1.44
Pillow==5.4.1
pluggy==0.13.0
psutil==5.5.1
py==1.8.0
pycrypto==2.6.1
pygame==1.9.4.post1
Pygments==2.3.1
PyGObject==3.30.4
pyinotify==0.9.6
PyJWT==1.7.0
pylint==2.2.2
pyOpenSSL==19.0.0
pyparsing==2.4.5
pyperclip==1.6.4
pyserial==3.4
pytest==5.3.0
python-apt==1.8.4
python-dateutil==2.8.1
pytz==2019.3
pyxdg==0.25
rainbowhat==0.1.0
requests==2.21.0
requests-oauthlib==1.0.0
responses==0.9.0
roman==2.0.0
RPi.GPIO==0.7.0
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
sense-hat==2.2.0
simplejson==3.16.0
six==1.12.0
skywriter==0.0.7
sn3218==1.2.7
snowballstemmer==2.0.0
soupsieve==1.8
Sphinx==2.2.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spidev==3.3
ssh-import-id==5.7
thonny==3.1.0
thonny-pi==1.1
touchphat==0.0.1
twython==3.7.0
typed-ast==1.3.1
unicornhathd==0.0.4
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
zipp==0.6.0
